Update to use new service module term
