--- Runtime/CameraSystem/Providers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6a49bf4e1c001054abbefd45e215ef03
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
