+++ Runtime/InputService/Profiles/PicoControllerServiceModuleProfile.cs
@@ -3,24 +3,19 @@
 
 using RealityCollective.Definitions.Utilities;
 using RealityToolkit.Definitions.Controllers;
-using RealityToolkit.Pico.InputSystem.Controllers;
 
-namespace RealityToolkit.Pico.InputSystem.Profiles
+namespace RealityToolkit.Pico.InputService.Profiles
 {
     /// <summary>
     /// Configuration profile for <see cref="Providers.PicoControllerDataProvider"/>.
     /// </summary>
-    public class PicoControllerDataProviderProfile : BaseMixedRealityControllerServiceModuleProfile
+    public class PicoControllerServiceModuleProfile : BaseMixedRealityControllerServiceModuleProfile
     {
         /// <inheritdoc />
         public override ControllerDefinition[] GetDefaultControllerOptions()
         {
             return new[]
             {
-                new ControllerDefinition(typeof(PicoG24KController), Handedness.Left),
-                new ControllerDefinition(typeof(PicoG24KController), Handedness.Right),
-                new ControllerDefinition(typeof(PicoNeo2Controller), Handedness.Left),
-                new ControllerDefinition(typeof(PicoNeo2Controller), Handedness.Right),
                 new ControllerDefinition(typeof(PicoNeo3Controller), Handedness.Left),
                 new ControllerDefinition(typeof(PicoNeo3Controller), Handedness.Right),
                 new ControllerDefinition(typeof(Pico4Controller), Handedness.Left),
