--- Runtime/InputSystem/Controllers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 86138a69ec4de1445b7d91743c8303a4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
