--- Runtime/InputSystem/Controllers/PicoG24KController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cdca989eb12104645bef2f1e5dafbf42
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
