--- Runtime/InputSystem/Controllers/PicoNeo2Controller.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ad4c4b82b76a6034b96f33c55b1733b1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
