--- Runtime/InputSystem/Providers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 04c604bb5880cb04c8368d78d46e76f7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
